+++ packages/core/src/types/PluginFunctions.ts
@@ -6,6 +6,7 @@
  * @copyright 2019-2022 Nguyen Huu Phuoc <me@phuoc.ng>
  */
 
+import { RotateDirection } from '../structs/RotateDirection';
 import { ScrollMode } from '../structs/ScrollMode';
 import { SpecialZoomLevel } from '../structs/SpecialZoomLevel';
 import type { ViewerState } from './ViewerState';
@@ -23,8 +24,8 @@ export interface PluginFunctions {
     ): void;
     jumpToPage(pageIndex: number): void;
     openFile(file: File): void;
-    rotate(rotation: number): void;
-    rotatePage(pageIndex: number, rotation: number): void;
+    rotate(direction: RotateDirection): void;
+    rotatePage(pageIndex: number, direction: RotateDirection): void;
     setViewerState(viewerState: ViewerState): void;
     switchScrollMode(scrollMode: ScrollMode): void;
     zoom(scale: number | SpecialZoomLevel): void;