+++ packages/rotate/src/index.d.ts
@@ -7,6 +7,7 @@
  */
 
 import * as React from 'react';
+import { RotateDirection } from '@react-pdf-viewer/core';
 import type { Plugin } from '@react-pdf-viewer/core';
 
 // Types
@@ -29,12 +30,6 @@ export interface RenderRotateProps {
 
 export interface RotateDecoratorProps {
     onClick(): void;
-}
-
-// Structs
-export enum RotateDirection {
-    Backward = 'Backward',
-    Forward = 'Forward',
 }
 
 // Plugin