+++ CHANGELOG.md
@@ -9,7 +9,8 @@
 
 ```tsx
 <Viewer
-    onRotate={({ doc, rotation }) => {
+    onRotate={({ direction, doc, rotation }) => {
+        // `direction` is the rotate direction
         // `doc` is the current document
         // `rotation` is the latest rotation value
     }}
@@ -20,10 +21,10 @@
 
 ```tsx
 const renderThumbnailItem = (props: RenderThumbnailItemProps) => (
-    <MinimalButton onClick={() => props.onRotatePage(90)}>
+    <MinimalButton onClick={() => props.onRotatePage(RotateDirection.Forward)}>
         <RotateForwardIcon />
     </MinimalButton>
-    <MinimalButton onClick={() => props.onRotatePage(-90)}>
+    <MinimalButton onClick={() => props.onRotatePage(RotateDirection.Backward)}>
         <RotateBackwardIcon />
     </MinimalButton>
 );
@@ -41,10 +42,10 @@ const renderPage: RenderPage = (props: RenderPageProps) => (
     <>
         {props.canvasLayer.children}
         <div>
-            <MinimalButton onClick={() => props.rotatePage(90)}>
+            <MinimalButton onClick={() => props.onRotatePage(RotateDirection.Forward)}>
                 <RotateForwardIcon />
             </MinimalButton>
-            <MinimalButton onClick={() => props.rotatePage(-90)}>
+            <MinimalButton onClick={() => props.onRotatePage(RotateDirection.Backward)}>
                 <RotateBackwardIcon />
             </MinimalButton>
         </div>
@@ -74,7 +75,8 @@ const { RotatePage } = rotatePluginInstance;
 
 ```tsx
 <Viewer
-    onRotate={({ doc, pageIndex, rotation }) => {
+    onRotatePage={({ direction, doc, pageIndex, rotation }) => {
+        // `direction` is the rotate direction
         // `doc` is the current document
         // `pageIndex` is the zero-based page index
         // `rotation` is the latest rotation value
@@ -93,6 +95,30 @@ const { RotatePage } = rotatePluginInstance;
 -   Typo in full screen change event
 -   There is a visible page that isn't rendered when setting the zoom level as page width
 -   The thumbnails aren't rotated after rotating the document
+
+**Breaking changes**
+
+-   The `RotateDirection` provided by the `@react-pdf-viewer/rotate` package now belongs to the `@react-pdf-viewer/core` package:
+
+```js
+// v3.1.2 and previous versions
+import { RotateDirection } from '@react-pdf-viewer/rotate';
+
+// From v3.2.0
+import { RotateDirection } from '@react-pdf-viewer/core';
+```
+
+-   The `rotate` function used in the plugins changes the parameter type:
+
+```js
+// v3.1.2 and previous versions
+rotate(90);
+rotate(-90);
+
+// From v3.2.0
+rotate(RotateDirection.Forward);
+rotate(RotateDirection.Backward);
+```
 
 ## v3.1.2